Split remote project integration tests
